+++ Common.Mod.Common/Config/IConfig.cs
@@ -3,5 +3,5 @@ namespace Common.Mod.Common.Config;
 public interface IConfig
 {
     public void Reset();
-    public void Render();
+    public void Render(IConfigUi configUi);
 }